requirements.txt
streamlit
pdfplumber
fpdf2
duckduckgo_search
pandas
